Move model helpers to their own module
